Move type check blk to file
