+++ src/types/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod blk;
 pub mod type_checker;
