+++ src/types/src/type_checker.rs
@@ -1,4 +1,4 @@
-use parser::structs::*;
+use parser::{structs::*, Parser};
 use std::{collections::HashMap, fmt::Display};
 
 use parser::structs::{BlockSeq, Decl, Expr, Type};
@@ -6,7 +6,7 @@ use parser::structs::{BlockSeq, Decl, Expr, Type};
 #[derive(Debug, PartialEq)]
 pub struct TypeErrors {
     errs: Vec<String>,
-    cont: bool,
+    pub(crate) cont: bool,
 }
 
 impl TypeErrors {
@@ -69,7 +69,7 @@ pub fn new_env_with_syms(syms: Vec<String>) -> Env {
 /// Struct to enable type checking by encapsulating type environment.
 pub struct TypeChecker<'prog> {
     program: &'prog BlockSeq,
-    envs: Vec<Env>,
+    pub(crate) envs: Vec<Env>,
 }
 
 impl<'prog> TypeChecker<'prog> {
@@ -161,7 +161,7 @@ impl<'prog> TypeChecker<'prog> {
 
     /// Return the type errors out instead of using mutable ref
     // because for nested errors in the expr we want to propagate those
-    fn check_expr(&mut self, expr: &Expr) -> Result<Type, TypeErrors> {
+    pub(crate) fn check_expr(&mut self, expr: &Expr) -> Result<Type, TypeErrors> {
         let local_errs = TypeErrors::new();
         let ty = match expr {
             Expr::Integer(_) => Type::Int,
@@ -240,7 +240,7 @@ impl<'prog> TypeChecker<'prog> {
     }
 
     /// Type check declaration and add errors if any
-    fn check_decl(&mut self, decl: &Decl) -> Result<(), TypeErrors> {
+    pub(crate) fn check_decl(&mut self, decl: &Decl) -> Result<(), TypeErrors> {
         // dbg!("Type checking decl:", decl);
         match decl {
             Decl::LetStmt(stmt) => {
@@ -271,52 +271,52 @@ impl<'prog> TypeChecker<'prog> {
         Ok(())
     }
 
-    fn check_block(&mut self, program: &BlockSeq) -> Result<Type, TypeErrors> {
-        let mut errs = TypeErrors::new();
-        // map bindings to types
-        // let mut ty_env: HashMap<String, Type> = HashMap::new();
-        // let mut ty_env = TyEnv::new();
-        let env = new_env_with_syms(program.symbols.clone());
-        self.envs.push(env);
+    // fn check_block(&mut self, program: &BlockSeq) -> Result<Type, TypeErrors> {
+    //     let mut errs = TypeErrors::new();
+    //     // map bindings to types
+    //     // let mut ty_env: HashMap<String, Type> = HashMap::new();
+    //     // let mut ty_env = TyEnv::new();
+    //     let env = new_env_with_syms(program.symbols.clone());
+    //     self.envs.push(env);
 
-        for decl in program.decls.iter() {
-            if let Err(mut decl_errs) = self.check_decl(decl) {
-                errs.append(&mut decl_errs);
+    //     for decl in program.decls.iter() {
+    //         if let Err(mut decl_errs) = self.check_decl(decl) {
+    //             errs.append(&mut decl_errs);
 
-                // if this err means we can't proceed, stop e.g let x = -true; let y = x + 3; - we don't know type of x since invalid
-                if !decl_errs.cont {
-                    break;
-                }
-            }
-        }
+    //             // if this err means we can't proceed, stop e.g let x = -true; let y = x + 3; - we don't know type of x since invalid
+    //             if !decl_errs.cont {
+    //                 break;
+    //             }
+    //         }
+    //     }
 
-        // return errors for decls first if any, without checking expr
-        // because expr may be dependent
-        if !errs.is_ok() {
-            self.envs.pop();
-            return Err(errs);
-        }
+    //     // return errors for decls first if any, without checking expr
+    //     // because expr may be dependent
+    //     if !errs.is_ok() {
+    //         self.envs.pop();
+    //         return Err(errs);
+    //     }
 
-        // Return type of last expr if any. If errs, add to err list
-        if let Some(last) = &program.last_expr {
-            let res = self.check_expr(last);
-            match res {
-                Ok(ty) => {
-                    self.envs.pop();
-                    return Ok(ty);
-                }
-                Err(mut expr_errs) => errs.append(&mut expr_errs),
-            };
-        }
+    //     // Return type of last expr if any. If errs, add to err list
+    //     if let Some(last) = &program.last_expr {
+    //         let res = self.check_expr(last);
+    //         match res {
+    //             Ok(ty) => {
+    //                 self.envs.pop();
+    //                 return Ok(ty);
+    //             }
+    //             Err(mut expr_errs) => errs.append(&mut expr_errs),
+    //         };
+    //     }
 
-        self.envs.pop();
+    //     self.envs.pop();
 
-        if errs.is_ok() {
-            Ok(Type::Unit)
-        } else {
-            Err(errs)
-        }
-    }
+    //     if errs.is_ok() {
+    //         Ok(Type::Unit)
+    //     } else {
+    //         Err(errs)
+    //     }
+    // }
 
     pub fn type_check(mut self) -> Result<Type, TypeErrors> {
         self.check_block(self.program)
@@ -329,34 +329,32 @@ impl Default for TypeErrors {
     }
 }
 
+pub fn expect_pass(inp: &str, exp_type: Type) {
+    let prog = Parser::new_from_string(inp).parse().expect("Should parse");
+    let ty = TypeChecker::new(&prog).type_check();
+    assert_eq!(Ok(exp_type), ty)
+}
+
+// contains true means check if input contains exp_err. else check full equals
+pub fn expect_err(inp: &str, exp_err: &str, contains: bool) {
+    let prog = Parser::new_from_string(inp).parse().expect("Should parse");
+    let ty_err = TypeChecker::new(&prog)
+        .type_check()
+        .expect_err("Should err");
+
+    if contains {
+        dbg!(ty_err.to_string());
+        assert!(ty_err.to_string().contains(exp_err))
+    } else {
+        dbg!(ty_err.to_string());
+        assert_eq!(ty_err.to_string(), exp_err)
+    }
+}
+
 #[cfg(test)]
 mod tests {
+    use super::{expect_err, expect_pass};
     use parser::structs::Type;
-    use parser::Parser;
-
-    use super::TypeChecker;
-
-    fn expect_pass(inp: &str, exp_type: Type) {
-        let prog = Parser::new_from_string(inp).parse().expect("Should parse");
-        let ty = TypeChecker::new(&prog).type_check();
-        assert_eq!(Ok(exp_type), ty)
-    }
-
-    // contains true means check if input contains exp_err. else check full equals
-    fn expect_err(inp: &str, exp_err: &str, contains: bool) {
-        let prog = Parser::new_from_string(inp).parse().expect("Should parse");
-        let ty_err = TypeChecker::new(&prog)
-            .type_check()
-            .expect_err("Should err");
-
-        if contains {
-            dbg!(ty_err.to_string());
-            assert!(ty_err.to_string().contains(exp_err))
-        } else {
-            dbg!(ty_err.to_string());
-            assert_eq!(ty_err.to_string(), exp_err)
-        }
-    }
 
     #[test]
     fn test_type_check_basic() {
@@ -514,151 +512,151 @@ mod tests {
         );
     }
 
-    #[test]
-    fn test_type_check_blk_simple() {
-        let t = "{ 2 }";
-        expect_pass(t, Type::Int);
+    // #[test]
+    // fn test_type_check_blk_simple() {
+    //     let t = "{ 2 }";
+    //     expect_pass(t, Type::Int);
 
-        let t = "{ 2; true }";
-        expect_pass(t, Type::Bool);
+    //     let t = "{ 2; true }";
+    //     expect_pass(t, Type::Bool);
 
-        let t = "{ let x : float = 2.4; x }";
-        expect_pass(t, Type::Float);
+    //     let t = "{ let x : float = 2.4; x }";
+    //     expect_pass(t, Type::Float);
 
-        let t = "{ let x = 2.4; x; }";
-        expect_pass(t, Type::Unit);
+    //     let t = "{ let x = 2.4; x; }";
+    //     expect_pass(t, Type::Unit);
 
-        let t = "let y = { let x = true; x }; y";
-        expect_pass(t, Type::Bool);
+    //     let t = "let y = { let x = true; x }; y";
+    //     expect_pass(t, Type::Bool);
 
-        let t = "let y : int = { let x = true; x }; y";
-        expect_err(t, "has declared type int but assigned type bool", true);
-    }
+    //     let t = "let y : int = { let x = true; x }; y";
+    //     expect_err(t, "has declared type int but assigned type bool", true);
+    // }
 
-    #[test]
-    fn test_type_check_blk_scope() {
-        let t = r"
-        let x : int = 2;
-        {
-            let y : int = 3;
-        }
-        y
-        ";
-        expect_err(t, "'y' not declared", true);
+    // #[test]
+    // fn test_type_check_blk_scope() {
+    //     let t = r"
+    //     let x : int = 2;
+    //     {
+    //         let y : int = 3;
+    //     }
+    //     y
+    //     ";
+    //     expect_err(t, "'y' not declared", true);
 
-        let t = r"
-        let x : int = 2;
-        {
-            let x : bool = true;
-            let y : bool = x;
-        }
-        x
-        ";
-        expect_pass(t, Type::Int);
+    //     let t = r"
+    //     let x : int = 2;
+    //     {
+    //         let x : bool = true;
+    //         let y : bool = x;
+    //     }
+    //     x
+    //     ";
+    //     expect_pass(t, Type::Int);
 
-        let t = r"
-        let x : int = 2;
-        {
-            let x : bool = true;
-            let y : int = x;
-        }
-        x
-        ";
-        expect_err(t, "has declared type int but assigned type bool", true);
+    //     let t = r"
+    //     let x : int = 2;
+    //     {
+    //         let x : bool = true;
+    //         let y : int = x;
+    //     }
+    //     x
+    //     ";
+    //     expect_err(t, "has declared type int but assigned type bool", true);
 
-        let t = r"
-        let x : int = 2;
-        let z = {
-            let x : bool = true;
-            let y : bool = x;
-            y
-        };
-        x;
-        z
-        ";
-        expect_pass(t, Type::Bool);
-    }
+    //     let t = r"
+    //     let x : int = 2;
+    //     let z = {
+    //         let x : bool = true;
+    //         let y : bool = x;
+    //         y
+    //     };
+    //     x;
+    //     z
+    //     ";
+    //     expect_pass(t, Type::Bool);
+    // }
 
-    #[test]
-    fn test_type_check_blk_more() {
-        let t = r"
-        let x = 2; 
-        let y = 0; 
-        { 
-            let x = 3; 
-            y = 4; 
-        } 
-        x+y
-        ";
+    // #[test]
+    // fn test_type_check_blk_more() {
+    //     let t = r"
+    //     let x = 2;
+    //     let y = 0;
+    //     {
+    //         let x = 3;
+    //         y = 4;
+    //     }
+    //     x+y
+    //     ";
 
-        expect_pass(t, Type::Int);
+    //     expect_pass(t, Type::Int);
 
-        // gets type from parent scope correctly during assign
-        let t = r"
-        let x = 2; 
-        let y : bool = true;
-        { 
-            let x = 3; 
-            y = 4; 
-        } 
-        x+y
-        ";
+    //     // gets type from parent scope correctly during assign
+    //     let t = r"
+    //     let x = 2;
+    //     let y : bool = true;
+    //     {
+    //         let x = 3;
+    //         y = 4;
+    //     }
+    //     x+y
+    //     ";
 
-        expect_err(t, "'y' declared with type bool but assigned type int", true);
+    //     expect_err(t, "'y' declared with type bool but assigned type int", true);
 
-        // doesn't matter that shadowed var has diff type
-        let t = r"
-        let x : int = 20; 
-        let y = 0; 
-        let z : bool = { 
-            let x : bool = true; 
-            y = 4; 
-            x
-        };
-        x+y
-        ";
-        expect_pass(t, Type::Int);
+    //     // doesn't matter that shadowed var has diff type
+    //     let t = r"
+    //     let x : int = 20;
+    //     let y = 0;
+    //     let z : bool = {
+    //         let x : bool = true;
+    //         y = 4;
+    //         x
+    //     };
+    //     x+y
+    //     ";
+    //     expect_pass(t, Type::Int);
 
-        // blk with no last expr has unit type
-        let t = r"
-        let x : int = 20; 
-        let y = 0; 
-        let z : bool = { 
-            let x : bool = true; 
-            y = 4; 
-            x;
-        };
-        x+y
-        ";
-        expect_err(t, "'z' has declared type bool but assigned type ()", true);
+    //     // blk with no last expr has unit type
+    //     let t = r"
+    //     let x : int = 20;
+    //     let y = 0;
+    //     let z : bool = {
+    //         let x : bool = true;
+    //         y = 4;
+    //         x;
+    //     };
+    //     x+y
+    //     ";
+    //     expect_err(t, "'z' has declared type bool but assigned type ()", true);
 
-        // x declared again in block, so the assign looks for closest decl of x which is Uninit at time of x = true;
-        let t = r"
-        let x = 10;
-        {
-            x = true;
-            let x = false;
-        };
-        ";
+    //     // x declared again in block, so the assign looks for closest decl of x which is Uninit at time of x = true;
+    //     let t = r"
+    //     let x = 10;
+    //     {
+    //         x = true;
+    //         let x = false;
+    //     };
+    //     ";
 
-        expect_err(
-            t,
-            "[TypeError]: Identifier 'x' assigned before declaration",
-            false,
-        );
-    }
+    //     expect_err(
+    //         t,
+    //         "[TypeError]: Identifier 'x' assigned before declaration",
+    //         false,
+    //     );
+    // }
 
-    #[test]
-    fn test_type_check_blk_errs() {
-        let t = r"
-        let x : int = true;
-        let y = 20;
-        {
-            let y : bool = 20;
-        }
-        x + false
-        ";
+    // #[test]
+    // fn test_type_check_blk_errs() {
+    //     let t = r"
+    //     let x : int = true;
+    //     let y = 20;
+    //     {
+    //         let y : bool = 20;
+    //     }
+    //     x + false
+    //     ";
 
-        expect_err(t, "[TypeError]: 'x' has declared type int but assigned type bool\n[TypeError]: 'y' has declared type bool but assigned type int", true);
-    }
+    //     expect_err(t, "[TypeError]: 'x' has declared type int but assigned type bool\n[TypeError]: 'y' has declared type bool but assigned type int", true);
+    // }
 }
